refactor(tapir-processing): :art: Abstracted PointCloudSlice and require custom Tapir implementation
